dead code cleanup
