--- src/WebJobs.Script.WebHost/App_Data/secrets/httptrigger-csharp.json
@@ -1,9 +1,0 @@
-{
-  "keys": [
-    {
-      "name": "default",
-      "value": "9884kkdlkkdf83ksld8589kflss90sll5kjjsyfjskqv",
-      "encrypted": false
-    }
-  ]
-}
