--- src/WebJobs.Script.WebHost/App_Data/secrets/httptrigger-powershell-modules.json
@@ -1,9 +1,0 @@
-{
-  "keys": [
-    {
-      "name": "default",
-      "value": "8CTs65hqBcX3DVddZOGkPoksSaIDRck9byv1ATWbqJuOb9h8MrVZzA==",
-      "encrypted": false
-    }
-  ]
-}
