--- src/WebJobs.Script/Description/LanguageInvokerInitializationResult.cs
@@ -1,17 +1,0 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the MIT License. See License.txt in the project root for license information.
-
-using System;
-using System.Collections.Generic;
-
-namespace Microsoft.Azure.WebJobs.Script.Description
-{
-    public class LanguageInvokerInitializationResult
-    {
-        [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
-        public Dictionary<string, string> LanguageServiceCapability { get; set; }
-
-        // TODO
-        // Add more language specific capabilities
-    }
-}
