--- src/WebJobs.Script/Description/LanguageInvokerMessagesEventArgs.cs
@@ -1,16 +1,0 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the MIT License. See License.txt in the project root for license information.
-
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Microsoft.Azure.WebJobs.Script.Description
-{
-    public class LanguageInvokerMessagesEventArgs : EventArgs
-    {
-        public string Message { get; set; }
-    }
-}
